// src/types/jobs.type.ts
export type Job = {
  id: string;
  title: string;
  company: string;
  companyUrl?: string;
  logo?: string;
  location: string;
  type: string;
  salary: string;
  tags?: string[];
  description: string;
  requirements?: string[];
  isFeatured?: boolean;
  categoryId?: number;
  category?: string;
  experience?: string;
  featured?: string;
  postedAt?: string;
  createdAt?: string;
  jobTitle?: string;
  jobCategory?: string;
  jobType?: string;
  jobLocation?: string;
  salaryRange?: string;
  jobDescription?: string;
  companyName?: string;
  companyWebsite?: string;
  email?: string;
  fullName?: string;
  password?: string;
};